it compiles again!  lol...

+++ src/solver/smtsolver.h
@@ -16,6 +16,7 @@
 #ifndef _STOKE_SRC_SOLVER_SMTSOLVER_H
 #define _STOKE_SRC_SOLVER_SMTSOLVER_H
 
+#include <map>
 #include <vector>
 
 #include "src/ext/cpputil/include/container/bit_vector.h"
@@ -57,7 +58,7 @@ public:
   /** Get the satisfying assignment for a bit from the model. */
   virtual bool get_model_bool(const std::string& var) = 0;
   /** Get the satisfying assignment for an array */
-  virtual std::map<cpputil::BitVector, cpputil::BitVector> get_model_array(const std::string& var, size_t key_bits, size_t value_bits) = 0;
+  virtual std::map<cpputil::BitVector, cpputil::BitVector> get_model_array(const std::string& var, uint16_t key_bits, uint16_t value_bits) = 0;
 
   /** Check if the last query trivvered an error. */
   virtual bool has_error() {
